use Trajectory to refactor integrate_vf_ivp
